+++ Trieu Quan Co.,Ltd/DAL/tblStockIssueDetails_DAL.cs
@@ -11,7 +11,7 @@ namespace DAL
     public class tblStockIssueDetails_DAL
     {
         tblStockIssueDetails_DTO p;
-        public tblStockIssueDetails_DAL(string ID, string ExportCode, string ProductCode, int Quantity, int UnitPrice, int TotalAmount)
+        public tblStockIssueDetails_DAL(string ID, string ExportCode, string ProductCode, decimal Quantity, decimal UnitPrice, decimal TotalAmount)
         {
             p = new tblStockIssueDetails_DTO(ID, ExportCode, ProductCode, Quantity, UnitPrice, TotalAmount);
         }
